Community ready

+++ src/main/java/com/start/demo/Services/Users/UserServices.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface UserServices {
     List<User> findAll();
-    User findById(int id);
+    User findById(Long id);
     User findByEmail(String email);
     boolean existsByEmail(String email);
     User saveUser(User user);
